requirements.txt
Django
djangorestframework
djongo
django-cors-headers
python-dotenv